refactor: remove shared/bridges folder
