+++ packages/shared/model/bridge/bnb-arbitrum.ts
