+++ packages/shared/model/bridge/ethereum-darwinia.ts
@@ -1,4 +1,4 @@
-import { BridgeConfig, ContractConfig, LockEventsStorage } from '../bridge';
+import { BridgeConfig, ContractConfig, LockEventsStorage } from './bridge';
 
 interface EthereumDarwiniaContractConfig extends ContractConfig {
   fee: string; // e2d cross chain fee querying address
