+++ packages/shared/model/bridge/index.ts
@@ -1,2 +1,15 @@
+export * from './bnb-arbitrum';
 export * from './bridge';
+export * from './crabDVM-ethereum';
+export * from './crabDVM-heco';
+export * from './crabDVM-polygon';
 export * from './cross-chain';
+export * from './ethereum-darwinia';
+export * from './ethereum-heco';
+export * from './ethereum-polygon';
+export * from './heco-polygon';
+export * from './parachain-substrate';
+export * from './substrate-dvm';
+export * from './substrate-substrateDVM';
+export * from './substrateDVM-substrateDVM';
+export * from './supports';
