+++ packages/shared/model/bridge/substrate-dvm.ts
@@ -1,3 +1,3 @@
-import { BridgeConfig } from '../bridge/bridge';
+import { BridgeConfig } from './bridge';
 
 export type SubstrateDVMBridgeConfig = Required<Omit<BridgeConfig, 'contracts'>>;
