+++ packages/shared/model/bridge/supports.ts
@@ -1,0 +1,13 @@
+export type BridgeName =
+  | 'substrate-DVM'
+  | 'substrate-substrateDVM'
+  | 'ethereum-darwinia'
+  | 'parachain-substrate'
+  | 'crabDVM-heco'
+  | 'crabDVM-ethereum'
+  | 'ethereum-heco'
+  | 'ethereum-polygon'
+  | 'crabDVM-polygon'
+  | 'substrateDVM-substrateDVM'
+  | 'bnb-arbitrum'
+  | 'heco-polygon';
