--- packages/shared/model/bridges/index.ts
@@ -1,12 +1,0 @@
-export * from './crabDVM-ethereum';
-export * from './crabDVM-heco';
-export * from './crabDVM-polygon';
-export * from './ethereum-darwinia';
-export * from './parachain-substrate';
-export * from './substrate-dvm';
-export * from './substrate-substrateDVM';
-export * from './ethereum-heco';
-export * from './ethereum-polygon';
-export * from './substrateDVM-substrateDVM';
-export * from './heco-polygon';
-export * from './bnb-arbitrum';
